requirements.txt
mysql_connector_repackaged==0.3.1
python-decouple==3.4
